% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Lab 5 Function: Binomial experiment simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{number of experiments to perform}

\item{n}{number of trials to conduct per experiment}

\item{p}{probability of success}
}
\value{
prints a barplot and success table
}
\description{
Lab 5 Function: Binomial experiment simulation
}
\examples{
mybin()
}
