% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Lab 9 Bootstrap Confidence Interval Histogram}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{data set}

\item{fun}{function to calculate}

\item{alpha}{(1-alpha)*100 percent confidence interval}

\item{cx}{character expansion factor for text in histogram}

\item{...}{extra parameters for histogram if want to use}
}
\value{
histogram of bootstrap confidence interval and list with important variables
}
\description{
Lab 9 Bootstrap Confidence Interval Histogram
}
\examples{
myboot2(x=c(1,1,1,2,2,2,2,3,3,4,4,4))
}
