% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Lab 11 Confidence Interval for mu given single sample x}
\usage{
myci(x)
}
\arguments{
\item{x}{single sample}
}
\value{
95% confidence interval for mu
}
\description{
Lab 11 Confidence Interval for mu given single sample x
}
\examples{
myci(x = rnorm(30,mean=10,sd=12))
}
