% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Lab 8 Central Limit Theorem for Binomial Distribution}
\usage{
mycltb(n, iter, p = 0.5)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iteratiions}

\item{p}{probability of success on a single trial}
}
\value{
histogram of sample means distribution
}
\description{
Lab 8 Central Limit Theorem for Binomial Distribution
}
\examples{
mycltb(n=10,iter=10000,p=0.3)
}
