% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Piecewise regression}
\usage{
myf(x, xk, coef)
}
\arguments{
\item{x}{independent variable}

\item{xk}{knot point}

\item{coef}{vector of coefficients}
}
\value{
value of piecewise linear regression at value x
}
\description{
Piecewise regression
}
\examples{
myf(0, 18, coef=c(6.2188237,0.6741809,-0.5803229))
}
