% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Lab 6 Shaded Area Curve Function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean of distribution}

\item{sigma}{standard deviation of distribution}

\item{a}{probability value in P(X<=a)}
}
\value{
graph with shaded curve and list of mu, sigma, and probability
}
\description{
Lab 6 Shaded Area Curve Function
}
\examples{
myncurve(0,1,0)
}
