% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{Polynomial function}
\usage{
myplot(x)
}
\arguments{
\item{x}{a quantitative value}
}
\value{
a function in terms of x or f(x)
}
\description{
Polynomial function
}
\examples{
myplot(1)
}
