% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Project 1 Overbooking Problem to find Optimal Tickets Sold}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats in the flight}

\item{gamma}{probability of plane being overbooked}

\item{p}{probability of a 'show'}
}
\value{
named list and 2 plots giving optimal tickets sold (discrete and continuous)
}
\description{
Project 1 Overbooking Problem to find Optimal Tickets Sold
}
\examples{
ntickets(N=200, gamma=0.02, p=0.95)
}
